Releasing Logos2
After months of waiting for a response from @maciejhirsz, regarding the access to publishing to crates.io, I think it is better to move on and publish a new crate, logos2 (and related).

The libraries' name is the same, but crates have a `2` appended at the end. 

This will allow to community to continue working on Logos, and publishing new releases when possible.